+++ server/src/game/schema/GameState.ts
@@ -109,7 +109,6 @@ export class GameLevelSchema extends Schema {
   @type("string") state: GameLevelSchemaState = "warmup";
   @type("number") startingCountdown = 10;
   @type("number") remainingGameTimeMillis = -1;
-  @type("number") maxCatcherTimeMillis = 30000;
 
   @type({ map: BodySchema }) bodies = new MapSchema<BodySchema>();
   @type(TileMap) tileMap = new TileMap();